refactor(lb-components): Migrate 2D view worker to TypeScript
