+++ packages/lb-components/package.json
@@ -111,6 +111,7 @@
     "rollup": "^2.43.1",
     "rollup-plugin-esbuild": "^3.0.2",
     "rollup-plugin-postcss": "^4.0.2",
+    "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-web-worker-loader": "^1.6.1",
     "sass": "^1.53.0",
     "typescript": "^4.2.3"
