+++ packages/lb-components/tsconfig.json
@@ -32,6 +32,6 @@
     },
   },
   "include": [
-    "src/*", // 导入目录
+    "src", // 导入目录
   ]
 }
